chore: rename developer options tab to debug content
