+++ ui/pages/settings-v2/debug-tab/debug-content/debug-content.stories.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Debug from './index';
+
+const DebugStory = {
+  title: 'Pages/Settings/Debug',
+
+  component: Debug,
+};
+
+export const DefaultStory = ({ variant, address }) => <Debug />;
+
+DefaultStory.storyName = 'Default';
+
+export default DebugStory;
