+++ ui/pages/settings-v2/debug-tab/debug-content/debug-content.tsx
@@ -39,13 +39,13 @@ import { getEnvironmentType } from '../../../../../app/scripts/lib/util';
 import { ENVIRONMENT_TYPE_POPUP } from '../../../../../shared/constants/app';
 import { getRemoteFeatureFlags } from '../../../../selectors';
 import { ConfirmationsDeveloperOptions } from '../../../confirmations/components/developer/confirmations-developer-options';
-import ToggleRow from './developer-options-toggle-row-component';
+import ToggleRow from './toggle-row-component';
 import SentryTest from './sentry-test';
 import { BackupAndSyncDevSettings } from './backup-and-sync';
 import MigrateToSplitStateTest from './migrate-to-split-state-test';
 
 /**
- * Settings Page for Developer Options (internal-only)
+ * Content for Debug Tab (internal-only)
  *
  * This page does not need i18n translation support because it's an internal settings page.
  * We only support the t('developerOptions') translation because the general settings architecture
@@ -53,7 +53,7 @@ import MigrateToSplitStateTest from './migrate-to-split-state-test';
  *
  * @returns
  */
-const DeveloperOptionsTab = () => {
+const DebugContent = () => {
   const t = useI18nContext();
   const dispatch = useDispatch();
   const navigate = useNavigate();
@@ -306,4 +306,4 @@ const DeveloperOptionsTab = () => {
   );
 };
 
-export default DeveloperOptionsTab;
+export default DebugContent;
