+++ ui/pages/settings-v2/debug-tab/debug-content/index.scss
