+++ ui/pages/settings-v2/debug-tab/debug-content/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './debug-content';
