+++ ui/pages/settings-v2/debug-tab/debug-content/sentry-test.tsx
