+++ ui/pages/settings-v2/debug-tab/debug-content/toggle-row-component.tsx
