--- ui/pages/settings-v2/debug-tab/developer-options-tab/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './developer-options-tab';
