+++ ui/pages/settings/index.scss
@@ -1,6 +1,6 @@
 @use "design-system";
 
-@import '../settings-v2/debug-tab/developer-options-tab/index';
+@import '../settings-v2/debug-tab/debug-content/index';
 @import 'networks-tab/index';
 @import 'transaction-shield-tab/index';
